Crate alloc_compose

Composable allocator structures for plugging together more powerful allocators.

alloc-compose currently uses the alloc-wg crate as backend. As soon as all features has landed upstream, this dependency will be dropped. Until AllocRef has been stabilized, this crate requires a nightly compiler.

The design of composable allocators is inspired by std::allocator Is to Allocation what std::vector Is to Vexation by Andrei Alexandrescu and the Phobos Standard Library of the D Programming Language.

Structs

FallbackAlloc
An allocator equivalent of an "or" operator in algebra.

NullAlloc
An emphatically empty implementation of AllocRef.

SegregateAlloc
Dispatches calls to AllocRef between two allocators depending on the size allocated.

Traits

Owns
Trait to determine if a given MemoryBlock is owned by an allocator.