Struct alloc_compose::region::IntrusiveRegion

pub struct IntrusiveRegion<'mem> { /* fields omitted */ }

An intrusive region allocator, which stores the current posision in the provided memory.

It holds a lifetime to the provided memory block, which ensures, that the allocator does not outlive the underlying memory.

For a version without lifetime see RawIntrusiveRegion instead.

Implementations

impl<'mem> IntrusiveRegion<'mem>

pub fn new(memory: &'mem mut [MaybeUninit<u8>]) -> Self

Creates a new region from the given memory block.

Panics

This function panics, when memory is not large enough to properly store a pointer.