Module alloc_compose::stats

Structures to collect allocator statistics.

Please see the Proxy documentation for examples.

Structs

AtomicCounter
An atomic counter for collectiong statistics which can be shared between threads.

Counter
A primitive counter for collectiong statistics.

FilteredAtomicCounter
An atomic counter for collectiong and filtering statistics which can be shared between threads.

FilteredCounter
A counter for collectiong and filtering statistics.

Enums

AllocInitFilter

ReallocPlacementFilter

ResultFilter