Struct alloc_compose::Chunk

pub struct Chunk<A, const SIZE: usize>(pub A);

Allocate memory with a multiple size of the provided chunk size.

Examples

#![feature(allocator_api, slice_ptr_len)]

use alloc_compose::{Chunk, Region};
use std::{
    alloc::{AllocRef, Layout},
    mem::MaybeUninit,
};

let mut data = [MaybeUninit::new(0); 64];
let mut alloc = Chunk::<_, 64>(Region::new(&mut data));
let ptr = alloc.alloc(Layout::new::<[u8; 16]>())?;
assert_eq!(ptr.len() % 32, 0);
assert!(ptr.len() >= 32);

When growing or shrinking the memory, Chunk will try to alter the memory in place before delegating to the underlying allocator.

#![feature(slice_ptr_get)]

use alloc_compose::ReallocInPlace;

let len = unsafe { alloc.grow_in_place(ptr.as_non_null_ptr(), Layout::new::<[u8; 16]>(), 24)? };
assert_eq!(len % 32, 0);
assert!(len >= 32);