Struct alloc_compose::Null

pub struct Null;

An emphatically empty implementation of AllocRef.

Although it has no direct use, it is useful as a "terminator" in composite allocators or for disabling the global allocator.

Examples

The Null will always return Err:

#![feature(allocator_api)]

use alloc_compose::Null;
use std::alloc::{AllocRef, Global, Layout};

let memory = Null.alloc(Layout::new::<u32>());
assert!(memory.is_err())

Even if a zero-sized allocation is requested:

let memory = Null.alloc(Layout::new::<()>());
assert!(memory.is_err())

Disabling the global allocator

use alloc_compose::Null;

#[global_allocator]
static A: Null = Null;