Struct alloc_compose::Proxy

pub struct Proxy<A, C> {
    pub alloc: A,
    pub callbacks: C,
}

Calls the provided callbacks when invoking methods on AllocRef.

A typical use case for a Proxy allocator is collecting statistics. alloc-compose provides different implementations for CallbackRef.

Examples

#![feature(allocator_api, slice_ptr_get)]

use alloc_compose::{stats, CallbackRef, Proxy};
use std::alloc::{AllocRef, Layout, System};

let counter = stats::Counter::default();
let mut alloc = Proxy {
    alloc: System,
    callbacks: counter.by_ref(),
};

unsafe {
    let memory = alloc.alloc(Layout::new::<u32>())?;
    alloc.dealloc(memory.as_non_null_ptr(), Layout::new::<u32>());
}

assert_eq!(counter.num_allocs(), 1);
assert_eq!(counter.num_deallocs(), 1);

If more information is needed, one can either implement CallbackRef itself or use a more fine-grained callback:

use alloc_compose::{
    region::Region,
    stats::{AllocInitFilter, ResultFilter},
};
use core::mem::MaybeUninit;

let counter = stats::FilteredCounter::default();
let mut data = [MaybeUninit::new(0); 32];
let mut alloc = Proxy {
    alloc: Region::new(&mut data),
    callbacks: counter.by_ref(),
};

unsafe {
    let memory = alloc.alloc(Layout::new::<u32>())?;
    alloc.dealloc(memory.as_non_null_ptr(), Layout::new::<u32>());

    alloc.alloc_zeroed(Layout::new::<[u32; 64]>()).unwrap_err();
}

assert_eq!(counter.num_allocates(), 2);
assert_eq!(
    counter.num_allocates_filter(AllocInitFilter::None, ResultFilter::Ok),
    1
);
assert_eq!(
    counter.num_allocates_filter(AllocInitFilter::Zeroed, ResultFilter::Err),
    1
);

Fields
alloc: Acallbacks: C