Struct alloc_compose::Region

pub struct Region<'a> { /* fields omitted */ }

Allocator over an user-defined region of memory.

Examples

#![feature(allocator_api)]

use alloc_compose::{Owns, Region};
use core::alloc::{AllocInit, AllocRef, Layout};

let mut data = [0; 64];
let mut region = Region::new(&mut data);

let memory = region.alloc(Layout::new::<u32>(), AllocInit::Uninitialized)?;
assert!(region.owns(memory));

It's possible to deallocate the latest memory block allocated:

unsafe { region.dealloc(memory.ptr, Layout::new::<u32>()) };
assert!(!region.owns(memory));

Implementations

impl<'a> Region<'a>

pub fn new(data: &'a mut [u8]) -> Self

pub const fn capacity(&self) -> usize

Returns the total capacity available in this allocator.

pub fn capacity_left(&self) -> usize

Returns the free capacity left for allocating.

pub fn reset(&mut self)

Resets the allocator.

This implies, that capacity() and capacity_left() are equal afterwards.

pub fn is_last_block(&self, memory: MemoryBlock) -> bool

Checks if memory is the latest block, which was allocated. For those blocks, it's possible to deallocate them or to grow or shrink them in place.