Trait alloc_compose::AllocAll

pub trait AllocAll {
    fn alloc_all(
        &mut self, 
        layout: Layout, 
        init: AllocInit
    ) -> Result<MemoryBlock, AllocErr>;

    fn dealloc_all(&mut self);

    fn capacity(&self) -> usize;

    fn capacity_left(&self) -> usize;

    fn is_empty(&self) -> bool { ... }

    fn is_full(&self) -> bool { ... }
}

Required methods

fn alloc_all(
&mut self,
layout: Layout,
init: AllocInit
) -> Result<MemoryBlock, AllocErr>

Attempts to allocate all of the memory the allocator can provide.

If the allocator is currently not managing any memory, then it returns all the memory available to the allocator. Subsequent calls should not suceed.

On success, returns a MemoryBlock meeting the size and alignment guarantees of layout.

The returned block is at least as large as specified by layout.size() and is initialized as specified by init, all the way up to the returned size of the block.

Errors

Returning Err indicates that either memory is exhausted or layout does not meet allocator's size or alignment constraints.

Implementations are encouraged to return Err on memory exhaustion rather than panicking or aborting, but this is not a strict requirement. (Specifically: it is legal to implement this trait atop an underlying native allocation library that aborts on memory exhaustion.)

Clients wishing to abort computation in response to an allocation error are encouraged to call the handle_alloc_error function, rather than directly invoking panic! or similar.

fn dealloc_all(&mut self)

Deallocates all the memory the allocator had allocated.

fn capacity(&self) -> usize

Returns the total capacity available in this allocator.

fn capacity_left(&self) -> usize

Returns the free capacity left for allocating.

Loading content...
Provided methods

fn is_empty(&self) -> bool

Returns if the allocator is currently not holding memory.

fn is_full(&self) -> bool

Returns if the allocator has no more capacity left.

Loading content...