Struct alloc_compose::Chunk

pub struct Chunk<A, const SIZE: usize>(pub A);

Allocate memory with a multiple size of the provided chunk size.

Examples

#![feature(allocator_api)]

use alloc_compose::Chunk;
use std::alloc::{AllocInit, AllocRef, Layout, System};

let mut data = [0; 64];
let mut alloc = Chunk::<_, 64>(System);
let memory = alloc.alloc(Layout::new::<[u8; 16]>(), AllocInit::Uninitialized)?;
assert_eq!(memory.size % 32, 0);
assert!(memory.size >= 32);

When growing or shrinking the memory, Chunk will try to alter the memory in place before delegating to the underlying allocator.

use std::alloc::ReallocPlacement;
let memory = unsafe {
    alloc.grow(
        memory.ptr,
        Layout::new::<[u8; 16]>(),
        24,
        ReallocPlacement::InPlace,
        AllocInit::Uninitialized,
    )?
};
assert_eq!(memory.size % 32, 0);
assert!(memory.size >= 32);